extract useCollapsibleStages hook and add Sentry error reporting
- Extract shared collapse/expand logic from TheoryScreenBody and AuralScreenBody into useCollapsibleStages hook; eliminates duplicate state, refs, timer management, and useFocusEffect init pattern
- Remove refreshKey + useFocusEffect hack from AuralScreenBody in favour of hook's built-in useFocusEffect
- Remove stale inline comments from both screen bodies
- Install @sentry/react-native; init in production with EXPO_PUBLIC_SENTRY_DSN; route console.error through Sentry.captureException instead of silencing it
