+++ app/_layout.tsx
@@ -1,3 +1,4 @@
+import * as Sentry from '@sentry/react-native'
 import { BottomSheetModalProvider } from '@gorhom/bottom-sheet'
 import { Stack } from 'expo-router'
 import { StatusBar } from 'expo-status-bar'
@@ -16,12 +17,19 @@ import { ThemeModeProvider } from '@/hooks/useThemeModeContext'
 import { UserProvider } from '@/hooks/useUserContext'
 import { SplashScreen } from '@/screens/SplashScreen'
 
+const SENTRY_DSN = process.env.EXPO_PUBLIC_SENTRY_DSN
+
 if (!__DEV__) {
+  if (SENTRY_DSN) {
+    Sentry.init({ dsn: SENTRY_DSN, environment: 'production' })
+  }
   console.log = () => {}
   console.warn = () => {}
-  console.error = () => {}
   console.info = () => {}
   console.debug = () => {}
+  console.error = SENTRY_DSN
+    ? (...args: unknown[]) => { Sentry.captureException(args[0] instanceof Error ? args[0] : new Error(String(args[0]))) }
+    : () => {}
 }
 
 const DevErrorBoundaryTrigger = () => {
