+++ package.json
@@ -29,6 +29,7 @@
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.8",
+    "@sentry/react-native": "~7.2.0",
     "babel-plugin-react-compiler": "^1.0.0",
     "expo": "~54.0.33",
     "expo-asset": "~12.0.12",
