+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useLastLesson } from '../screens/HomeScreen/components/LessonCard/useLastLesson'
+export { useCollapsibleStages } from './useCollapsibleStages'
 export { useDevice, useResponsiveValue } from './useDevice'
 export { useMetronome } from './useMetronome'
 export { usePlayer } from './usePlayer'
