+++ src/screens/TheoryScreen/TheoryScreenBody/index.tsx
@@ -1,8 +1,6 @@
-import { useProgress } from '@/hooks'
+import { useCollapsibleStages, useProgress } from '@/hooks'
 import type { Stage, StageLesson } from '@types'
-import { useFocusEffect } from 'expo-router'
-import { useCallback, useEffect, useRef, useState } from 'react'
-import { Animated, ScrollView, View } from 'react-native'
+import { useCallback } from 'react'
 import { LessonDivider } from '../components/LessonDivider'
 import { LessonSection } from '../components/LessonSection'
 import { StageHeader } from '../components/StageHeader'
@@ -13,8 +11,8 @@ const PREVIEW_LESSONS_COUNT = 2
 
 export const TheoryScreenBody = () => {
   const { getStageById, getStageRequirements, getNextLockedStage, stages } = useProgress()
-  const stagesRef = useRef(stages)
-  stagesRef.current = stages
+  const { scrollViewRef, collapsedStages, visibleStages, animatedHeights, stageRefs, toggleStageCollapse } =
+    useCollapsibleStages(stages)
 
   const getVisibleLessonsForStage = useCallback((stageId: string): StageLesson[] => {
     const stage = getStageById(stageId)
@@ -35,102 +33,9 @@ export const TheoryScreenBody = () => {
         !requirements.isUnlocked && requirements.progressNeeded.length > 0
           ? requirements.progressNeeded.join(', ')
           : ''
-      return {
-        stage,
-        isAccessible: requirements.isUnlocked,
-        blockingMessage,
-        lessons
-      }
+      return { stage, isAccessible: requirements.isUnlocked, blockingMessage, lessons }
     },
     [getStageById, getStageRequirements, getVisibleLessonsForStage]
-  )
-
-  const scrollViewRef = useRef<ScrollView>(null)
-  const [collapsedStages, setCollapsedStages] = useState<Record<string, boolean>>({})
-  const [visibleStages, setVisibleStages] = useState<Record<string, boolean>>({})
-  const animatedHeights = useRef<Record<string, Animated.Value>>({})
-  const stageRefs = useRef<Record<string, View>>({})
-  const openedStageIdsRef = useRef<Set<string>>(new Set())
-
-  const scrollToStage = useCallback((stageId: string, offset = 100) => {
-    const stageRef = stageRefs.current[stageId]
-    const scrollRef = scrollViewRef.current
-    if (!stageRef || !scrollRef) return
-    stageRef.measureLayout(
-      scrollRef as unknown as View,
-      (x: number, y: number) => {
-        scrollViewRef.current?.scrollTo({ y: Math.max(0, y - offset), animated: true })
-      },
-      () => scrollViewRef.current?.scrollToEnd({ animated: true })
-    )
-  }, [])
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      scrollViewRef.current?.scrollToEnd({ animated: true })
-    }, 100)
-
-    return () => clearTimeout(timer)
-  }, [])
-
-  const initCollapseState = useCallback(() => {
-    const currentStages = stagesRef.current
-    const initialCollapsedState: Record<string, boolean> = {}
-    const initialVisibleState: Record<string, boolean> = {}
-    const unlockedStages = currentStages.filter(s => s.isUnlocked)
-    const furthest = unlockedStages.length > 0
-      ? unlockedStages.reduce((max, s) => (s.order > max.order ? s : max), unlockedStages[0])
-      : undefined
-    const furthestId = furthest?.id
-
-    currentStages.forEach(stage => {
-      if (!animatedHeights.current[stage.id]) {
-        animatedHeights.current[stage.id] = new Animated.Value(1)
-      }
-
-      const shouldBeOpen = furthestId === stage.id
-
-      if (stage.isCleared) {
-        initialCollapsedState[stage.id] = !shouldBeOpen
-        initialVisibleState[stage.id] = shouldBeOpen
-        animatedHeights.current[stage.id].setValue(shouldBeOpen ? 1 : 0)
-      } else {
-        initialVisibleState[stage.id] = true
-        animatedHeights.current[stage.id].setValue(1)
-      }
-    })
-
-    setCollapsedStages(initialCollapsedState)
-    setVisibleStages(initialVisibleState)
-  }, [])
-
-  useFocusEffect(initCollapseState)
-
-  const toggleStageCollapse = useCallback(
-    (stageId: string) => {
-      const isCurrentlyCollapsed = collapsedStages[stageId]
-      const newCollapsedState = !isCurrentlyCollapsed
-
-      if (newCollapsedState) {
-        setVisibleStages(prev => ({ ...prev, [stageId]: false }))
-        setCollapsedStages(prev => ({ ...prev, [stageId]: true }))
-      } else {
-        setVisibleStages(prev => ({ ...prev, [stageId]: true }))
-        setCollapsedStages(prev => ({ ...prev, [stageId]: false }))
-        openedStageIdsRef.current.add(stageId)
-        setTimeout(() => {
-          animatedHeights.current[stageId].setValue(0)
-          Animated.timing(animatedHeights.current[stageId], {
-            toValue: 1,
-            duration: 200,
-            useNativeDriver: true
-          }).start()
-          setTimeout(() => scrollToStage(stageId, 0), 150)
-        }, 10)
-      }
-      if (newCollapsedState) openedStageIdsRef.current.delete(stageId)
-    },
-    [collapsedStages, scrollToStage]
   )
 
   const nextLockedStage = getNextLockedStage()
@@ -154,7 +59,7 @@ export const TheoryScreenBody = () => {
           }
           reserveLayoutSpace={!hasNextStage}
         />
-        
+
         {sortedStages.map((stage) => {
           const stageData = getStageDisplayData(stage.id)
           const isCollapsed = collapsedStages[stage.id] || false
@@ -166,12 +71,10 @@ export const TheoryScreenBody = () => {
           }
 
           return (
-            <StageContainer 
-              key={stage.id} 
+            <StageContainer
+              key={stage.id}
               ref={(ref) => {
-                if (ref) {
-                  stageRefs.current[stage.id] = ref
-                }
+                if (ref) stageRefs.current[stage.id] = ref
               }}
             >
               {stage.isCleared && lessonsToShow.length > 0 && (
@@ -184,29 +87,28 @@ export const TheoryScreenBody = () => {
               )}
               {isVisible && (
                 <CollapsibleLessonsContainer
-                  style={{
-                    opacity: stage.isCleared ? animatedHeights.current[stage.id] : 1
-                  }}
+                  style={{ opacity: stage.isCleared ? animatedHeights.current[stage.id] : 1 }}
                 >
                   {lessonsToShow
                     .slice()
                     .reverse()
                     .map((lesson, reversedIndex) => {
                       const originalIndex = lessonsToShow.length - 1 - reversedIndex
-                      const isLastLesson = reversedIndex === 0 // Topmost lesson in this stage
+                      const isLastLesson = reversedIndex === 0
                       const shouldBePartial = isPreviewStage && isLastLesson
-                      
                       return [
                         <PartialLessonContainer key={lesson.id} isPartial={shouldBePartial}>
                           <LessonContent isPartial={shouldBePartial}>
-                            <LessonSection 
-                              index={originalIndex} 
+                            <LessonSection
+                              index={originalIndex}
                               lesson={lesson}
                               allStageLessons={lessonsToShow}
                             />
                           </LessonContent>
                         </PartialLessonContainer>,
-                        reversedIndex < lessonsToShow.length - 1 && !lesson.isFinalTest && <LessonDivider key={`divider-${lesson.id}`} />
+                        reversedIndex < lessonsToShow.length - 1 && !lesson.isFinalTest && (
+                          <LessonDivider key={`divider-${lesson.id}`} />
+                        )
                       ]
                     })}
                 </CollapsibleLessonsContainer>
